Eliminate scenarios
We have phpunit load the correct autoloader and run composer in the individual integrations directories.
For PHP extensions, we build them as part of the container, then load via CLI -d extension= INI setting.
This is necessary to be able to run tests in parallel.

Also removing old, redundant OpenTracing1-betaX testsuites.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>